+++ pypeline/tools/bam_pipeline/mkfile.py
@@ -79,6 +79,9 @@ _TEMPLATE_BAM_OPTIONS = \
 
     # Settings for mappings performed using BWA
     BWA:
+      # One of "backtrack", "bwasw", or "mem"; see the BWA documentation
+      # for a description of each algorithm (defaults to 'backtrack')
+      Algorithm: backtrack
       # Filter hits with a mapping quality (Phred) below this value
       MinQuality: 0
       # Filter reads that did not map to the reference sequence